test(core test): improved core testing code

+++ hexonet/ispapicli/modules/scrap.py
@@ -1,4 +1,5 @@
 import sys
+from typing import Text
 from bs4 import BeautifulSoup
 from .db import DB
 from git import Repo
@@ -23,16 +24,18 @@ class Scrap:
         elif __file__:
             self.absolute_dirpath = os.path.dirname(__file__)
             self.repo_path = os.path.join(self.absolute_dirpath, "../APIDocsRepo/")
+        return True
 
     def __initCloneAPIDocsDir(self):
         if os.path.exists(self.repo_path):
             shutil.rmtree(self.repo_path)
         if not os.path.exists(self.repo_path):
             os.makedirs(self.repo_path)
+        return True
 
     def __saveCommandToDB(self, commandName, data):
         """
-        Creates a json file for the command: commandName
+        Creates a record in DB for the command: commandName
 
         Returns:
         --------
@@ -41,10 +44,10 @@ class Scrap:
         try:
             table = self.dbObj.db.table("commands")
             table.insert(data)
-            print("Command inserted: ", commandName)
+            print("Command saved: ", commandName)
             return True
         except Exception:
-            raise Exception("Couldn't create a file for the command: " + commandName)
+            raise Exception("Couldn't insert the command: " + commandName)
 
     def __getCommandData(self, commandName, description, availability, parameters):
         """
@@ -65,28 +68,26 @@ class Scrap:
             raise e
 
     def __cloneRepo(self):
-        self.__initCloneAPIDocsDir()
-        repo = Repo.clone_from(
-            "https://github.com/hexonet/hexonet-api-documentation.git",
-            self.repo_path,
-            branch="master",
-        )
-
-    def __pullRepo(self):
-        repo = Repo.git.pull_request(
-            "https://github.com/hexonet/hexonet-api-documentation.git",
-            self.repo_path,
-            branch="master",
-        )
+        if self.__initCloneAPIDocsDir():
+            repo = Repo.clone_from(
+                "https://github.com/hexonet/hexonet-api-documentation.git",
+                self.repo_path,
+                branch="master",
+            )
+            return True
+        return False
 
     def __getCommandsParams(self, raw):
+        # split the rows based on | delimiter
+        # this will return a list of paramname, min, type, etc.
         headers = raw[2].split("|")
         params = []
         param = {}
+        # skip uncessary spaces, start from line/row 4
         for i in range(4, len(raw)):
             if raw[i] != "----":
                 try:
-                    cols = raw[i].split("|")
+                    cols = raw[i].split("|")  # split each row data to match the headers
                     for j in range(0, len(cols)):
                         headType = headers[j].strip(" \t\n\r")
                         headValue = cols[j].strip(" \t\n\r")
@@ -101,7 +102,6 @@ class Scrap:
         return params
 
     # scrap commands
-
     def scrapCommands(self):
         """
         Executes the scrap process
@@ -111,43 +111,43 @@ class Scrap:
         Null
         """
         # clone repo
-        self.__cloneRepo()
-        # init database
-        self.dbObj.db.drop_table("commands")
-        commandName = ""
-        description = ""
-        availability = ""
-        parameters = []
-        # get all commands urls, ending with .md
-        mainDir = os.path.join(self.repo_path, "API")
-        result = list(Path(mainDir).glob("**/*.md"))
-        # section regex
-        secionRegex = r"^#{2}\s\w+$"
-        for dir in result:
-            # Load the file into file_content
-            file_content = []
-            for line in open(dir).readlines():
-                file_content.append(line.strip(" \t\n\r"))
-            # parse the content
-            for i in range(len(file_content)):
-                # first line is the command name
-                if i == 0:
-                    commandName = (file_content[i]).split(" ")[1]
-                # checkSection
-                section = re.findall(secionRegex, file_content[i])
-                if len(section) == 1:
-                    sectionName = (section[0].split(" "))[1]
-                    if sectionName == "DESCRIPTION":
-                        description = file_content[i + 1]
-                    if sectionName == "AVAILABILITY":
-                        availability = file_content[i + 1]
-                    if sectionName == "COMMAND":
-                        parameters = self.__getCommandsParams(file_content[i:])
-                        break
-            data = self.__getCommandData(
-                commandName, description, availability, parameters
-            )
-            self.__saveCommandToDB(commandName, data)
+        if self.__cloneRepo():
+            # init database | delete old database
+            self.dbObj.db.drop_table("commands")
+            commandName = ""
+            description = ""
+            availability = ""
+            parameters = []
+            # get all commands urls, ending with .md
+            mainDir = os.path.join(self.repo_path, "API")
+            result = list(Path(mainDir).glob("**/*.md"))
+            # section regex
+            secionRegex = r"^#{2}\s\w+$"
+            for dir in result:
+                # Load the file into file_content
+                file_content = []
+                for line in open(dir).readlines():
+                    file_content.append(line.strip(" \t\n\r"))
+                # parse the content
+                for i in range(len(file_content)):
+                    # first line is the command name
+                    if i == 0:
+                        commandName = (file_content[i]).split(" ")[1]
+                    # checkSection
+                    section = re.findall(secionRegex, file_content[i])
+                    if len(section) == 1:
+                        sectionName = (section[0].split(" "))[1]
+                        if sectionName == "DESCRIPTION":
+                            description = file_content[i + 1]
+                        if sectionName == "AVAILABILITY":
+                            availability = file_content[i + 1]
+                        if sectionName == "COMMAND":
+                            parameters = self.__getCommandsParams(file_content[i:])
+                            break
+                data = self.__getCommandData(
+                    commandName, description, availability, parameters
+                )
+                self.__saveCommandToDB(commandName, data)
 
-        print("\nCommands count: " + str(len(result)))
-        print("Command finished.")
+            print("\nCommands count: " + str(len(result)))
+            print("Command finished.")
